# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClearFunnel Hiring Governance Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rules:
    get:
      operationId: listRules
      tags: [rules]
      summary: List all filter rules with performance metrics
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived, pending_validation]
        - name: roleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Rules with performance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleListResponse"
    post:
      operationId: createRule
      tags: [rules]
      summary: Create a new filter rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilterRuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilterRule"
  /rules/{id}:
    get:
      operationId: getRule
      tags: [rules]
      summary: Get a filter rule with version history and validation runs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Filter rule with history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilterRuleDetail"
        "404":
          description: Not found
    patch:
      operationId: updateRule
      tags: [rules]
      summary: Update a filter rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilterRuleUpdate"
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilterRule"
    delete:
      operationId: archiveRule
      tags: [rules]
      summary: Archive a filter rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Archived
  /rules/{id}/validate:
    post:
      operationId: triggerRuleValidation
      tags: [rules]
      summary: Manually trigger a validation run for a rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Validation run started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationRun"

  # ─── Validation Runs ──────────────────────────────────────────
...
components:
  schemas:
...
    FilterRule:
      type: object
      required: [id, name, description, ruleType, criteria, status, version, createdBy, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        ruleType:
          type: string
          enum: [keyword_match, employment_gap, degree_requirement, experience_years, location, custom]
        criteria:
          type: string
        status:
          type: string
          enum: [active, archived, pending_validation]
        version:
          type: integer
        roleId:
          type: ["integer", "null"]
        roleName:
          type: ["string", "null"]
        lastValidatedAt:
          type: ["string", "null"]
          format: date-time
        validationStatus:
          type: ["string", "null"]
          enum: [passed, failed, not_run, null]
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        # Performance metrics
        triggeredCount:
          type: integer
        correctCount:
          type: integer
        falsePositiveCount:
          type: integer
        falseNegativeCount:
          type: integer
        weightPercent:
          type: integer
          description: Rule importance weight 0-100
        accuracyScore:
          type: number
          description: Percentage of triggered decisions that were correct
        fpRate:
          type: number
          description: False positive rate as a percentage
        fnRate:
          type: number
          description: False negative rate as a percentage
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        optimizerRecommendation:
          type: ["string", "null"]
          description: AI-generated suggestion to improve this rule
    RuleListResponse:
      type: object
      required: [rules, summary]
      properties:
        rules:
          type: array
          items:
            $ref: "#/components/schemas/FilterRule"
        summary:
          type: object
          properties:
            totalRules:
              type: integer
            activeRules:
              type: integer
            avgAccuracy:
              type: number
            totalTriggered:
              type: integer
            highRiskRules:
              type: integer
            rulesNeedingAttention:
              type: integer
    FilterRuleDetail:
      type: object
      required: [rule, history, validationRuns]
      properties:
        rule:
          $ref: "#/components/schemas/FilterRule"
        history:
          type: array
          items:
            $ref: "#/components/schemas/RuleHistoryEntry"
        validationRuns:
          type: array
          items:
            $ref: "#/components/schemas/ValidationRun"
    FilterRuleInput:
      type: object
      required: [name, description, ruleType, criteria, createdBy]
      properties:
        name:
          type: string
        description:
          type: string
        ruleType:
          type: string
        criteria:
          type: string
        roleId:
          type: integer
        createdBy:
          type: string
        weightPercent:
          type: integer
    FilterRuleUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        criteria:
          type: string
        status:
          type: string
        weightPercent:
          type: integer
        changedBy:
          type: string
    RuleHistoryEntry:
      type: object
      required: [version, changedBy, changes, changedAt]
      properties:
        version:
          type: integer
        changedBy:
          type: string
        changes:
          type: string
        changedAt:
          type: string
          format: date-time

    # ─── Validation ──────────────────────────────────────────────
    ValidationRun:
      type: object
      required: [id, ruleId, status, benchmarkCount, passCount, failCount, startedAt]
      properties:
        id:
          type: integer
        ruleId:
          type: integer
        ruleName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, running, passed, failed]
        benchmarkCount:
          type: integer
        passCount:
          type: integer
        failCount:
          type: integer
        passRate:
          type: number
          description: Pass percentage
        validationConfidence:
          type: integer
          description: 0-100 confidence in this validation result
        overrideJustification:
          type: ["string", "null"]
        overriddenBy:
          type: ["string", "null"]
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        results:
          type: array
          items:
            $ref: "#/components/schemas/ValidationResult"
...
    ValidationResult:
      type: object
      required: [benchmarkResumeId, candidateName, expectedOutcome, actualOutcome, passed]
      properties:
        benchmarkResumeId:
          type: integer
        candidateName:
          type: string
        expectedOutcome:
          type: string
          enum: [pass, fail]
        actualOutcome:
          type: string
          enum: [pass, fail]
        passed:
          type: boolean
        failReason:
          type: ["string", "null"]

    # ─── Benchmark Resumes ───────────────────────────────────────
...